% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_misclass_matrix_01.R
\name{admission}
\alias{admission}
\title{Accessory function to misclassification_01 that compares the cells of a trinarised matrix to a cluster's
representative}
\usage{
admission(trinarisedM2, representative1)
}
\description{
Accessory function to misclassification_01 that compares the cells of a trinarised matrix to a cluster's
representative
}
