% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DITASIC.R
\name{create.empdist}
\alias{create.empdist}
\title{Create empirical distribution adapted from DiTaSIC}
\usage{
create.empdist(abundance_objects, norm.fac = 1)
}
\description{
Create empirical distribution adapted from DiTaSIC
}
