% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_misclass_matrix_01.R
\name{create_misclass_matrix_01}
\alias{create_misclass_matrix_01}
\title{Main create misclassification function}
\usage{
create_misclass_matrix_01(bigM, clust)
}
\arguments{
\item{bigM}{A gene expression count matrix}

\item{clust}{A vector labelling each cell to a cluster}
}
\value{
A sqaure matrix of a dimension equal to the number of clusters present in the data
}
\description{
Main create misclassification function
}
