% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_misclass_matrix_02.R
\name{create_misclass_matrix_02}
\alias{create_misclass_matrix_02}
\title{Main function to create similarity matrix}
\usage{
create_misclass_matrix_02(bigM, clust)
}
\description{
Main function to create similarity matrix
}
