% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_misclass_matrix_01.R
\name{createRepresentatives}
\alias{createRepresentatives}
\title{Accessory function to misclassification_01 that creates representatives of a cluster}
\usage{
createRepresentatives(trinarisedM, clust)
}
\description{
Accessory function to misclassification_01 that creates representatives of a cluster
}
