% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcats_betabin.R
\name{dcats_betabin}
\alias{dcats_betabin}
\title{A beta-binomial regression test with similarity based bootstrapping}
\usage{
dcats_betabin(
  counts1,
  counts2,
  similarity_mat = NULL,
  n_samples = 50,
  pseudo_count = NULL
)
}
\arguments{
\item{counts1}{A matrix of compsition sizes (n_rep1, n_cluster) for each
replicate in each cluster for codition 1 as case}

\item{counts2}{A matrix of compsition sizes (n_rep2, n_cluster) for each
replicate in each cluster for codition 2 as control}

\item{similarity_mat}{A matrix of floats (n_cluster, n_cluster) for the
similarity matrix between cluster group pair. The order of cluster should
be consistent with those in `counts1` and `counts2`}

\item{n_samples}{An integer for number samples in sampling for estimating the
variance of the weights}

\item{pseudo_count}{A pseudo count to add for counts in all cell types.
Default NULL means 0 except if a cell type is emplty in one condition,
otherwise pseudo_count will be: 0.01 * rowMeans for each condition}
}
\value{
a vector of significance p values for each cluster
}
\description{
A GLM test with binomial distribution. In order to estimate the variance of
the weight, a boostrapping based on the composition similarity is performed.
}
\examples{
K <- 2
totals1 = c(100, 800, 1300, 600)
totals2 = c(250, 700, 1100)
diri_s1 = rep(1, K) * 20
diri_s2 = rep(1, K) * 20
simil_mat = create_simMat(K, confuse_rate=0.2)
sim_dat <- DCATS::simulator_base(totals1, totals2, diri_s1, diri_s2, simil_mat)
#dcats_betabin(sim_dat[[1]], sim_dat[[2]], simil_mat, n_samples = 100)

}
