% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcats_binom.R
\name{dcats_fit}
\alias{dcats_fit}
\title{A binomial regression test with similarity based bootstrapping}
\usage{
dcats_fit(counts1, counts2, similarity_mat = NULL, n_samples = 50)
}
\arguments{
\item{counts1}{A matrix of compsition sizes (n_rep1, n_cluster) for each
replicate in each cluster for codition 1}

\item{counts2}{A matrix of compsition sizes (n_rep2, n_cluster) for each
replicate in each cluster for codition 2}

\item{similarity_mat}{A matrix of floats (n_cluster, n_cluster) for the
similarity matrix between cluster group pair}
}
\value{
a vector of significance p values for each cluster
}
\description{
A GLM test with binomial distribution. In order to estimate the variance of
the weight, a boostrapping based on the composition similarity is performed.
}
\examples{
K <- 2
totals1 = c(100, 800, 1300, 600)
totals2 = c(250, 700, 1100)
diri_s1 = rep(1, K) * 20
diri_s2 = rep(1, K) * 20
confuse_rate = 0.2
simil_mat = diag(K) * (1 - confuse_rate) + confuse_rate * (1 - diag(K)) / 3
sim_dat <- DCATS::simulator_base(totals1, totals2, diri_s1, diri_s2, simil_mat)
dcats_fit(sim_dat[[1]], sim_dat[[2]], confuse_mat, n_samples = 100)

}
