% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcats_GLM.R
\name{dcats_GLM}
\alias{dcats_GLM}
\title{A Generalised linear model based likelihood ratio testing}
\usage{
dcats_GLM(
  count_mat,
  design_mat,
  similarity_mat = NULL,
  n_samples = 50,
  pseudo_count = NULL,
  base_model = "NULL"
)
}
\arguments{
\item{count_mat}{A matrix of compsition sizes (n_sample, n_cluster) for each
cluster in each sample}

\item{design_mat}{A matrix of testing candidate factors (n_sample, n_factor)
with same sample order as count_mat}

\item{base_model}{A string value: `NULL` for 1 factor vs NULL factor testing;
`FULL` for FULL facotrs vs n-1 factors testing. Only `NULL` is supported
now.}

\item{model}{A string value: `betabin` for beta-binomial and `negbin` for
negative binomial distribution}
}
\value{
a list of significance p values for each cluster
}
\description{
GLM supports both beta-binomial and negative binomial from aod package.
}
\examples{
K <- 3
totals1 = c(100, 800, 1300, 600)
totals2 = c(250, 700, 1100)
diri_s1 = rep(1, K) * 20
diri_s2 = rep(1, K) * 20
simil_mat = get_similarity_mat(K, confuse_rate=0.2)
sim_dat <- DCATS::simulator_base(totals1, totals2, diri_s1, diri_s2, simil_mat)
sim_count = rbind(sim_dat$numb_cond1, sim_dat$numb_cond2)
sim_design = matrix(c("g1", "g1", "g1", "g1", "g2", "g2", "g2"), ncol = 1)
dcats_GLM(sim_count, sim_design, similarity_mat = simil_mat)

}
