% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_reference.R
\name{detect_reference}
\alias{detect_reference}
\title{Calculate a global phi for all cell types}
\usage{
detect_reference(count_mat, design_mat, similarity_mat = NULL, fix_phi = NULL)
}
\arguments{
\item{count_mat}{A matrix of composition sizes (n_sample, n_cluster) for each
cluster in each sample.}

\item{design_mat}{A matrix or a data frame of testing candidate factors 
(n_sample, n_factor) with same sample order as count_mat. All factors 
should be continous and categorical with only two levels.}

\item{similarity_mat}{A matrix of floats (n_cluster, n_cluster) for the
similarity matrix between cluster group pair. The order of cluster should
be consistent with those in `count_mat`.}

\item{fix_phi}{A numeric used to provided a fixed phi value for the GLM 
for all cell types.}
}
\value{
A data frame with ordered cell types and their p-value. Cell types 
are ordered by their p-values. The order indicating how they are recommended
to be selected as reference cell types.
}
\description{
Assuming all cell types share the same phi. This global phi can be calculate 
by pooling all cell types together to fit a beta binomial distribution.
}
