% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DITASIC.R
\name{DiffExpr}
\alias{DiffExpr}
\title{Differential abundance analysis main function adapted from DiTASIC}
\usage{
DiffExpr(
  empdist1,
  empdist2,
  norm.coeff_s1,
  norm.coeff_s2,
  threshold.minreads = 0,
  seed = 1448
)
}
\description{
Differential abundance analysis main function adapted from DiTASIC
}
