% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DITASIC.R
\name{EstimateCount}
\alias{EstimateCount}
\title{Estimate count function adapted from DiTaSIC}
\usage{
EstimateCount(
  mat,
  observedCounts,
  filtering = FALSE,
  pvalThreshold = 0.05,
  minThreshold = 0
)
}
\arguments{
\item{mat}{A matrix of floats (n_group, n_group) for similarity matrix
between clusters}

\item{observedCounts}{A vector of integers (n_group) for the observed sample
count in each group}

\item{filtering}{A bool value for ...}

\item{pvalThreshold}{A float for ...}
}
\value{
a list containing \code{glm.coeffs}, a vector of estimated
coefficients; \code{error.abs}, a vector of standard divation of estimated
coefficients; \code{mod.exclude}, an indictor verctor of a cluster is
excluded; \code{raw.pval}, a vector of p values for each coefficient.
}
\description{
Estimate count function adapted from DiTaSIC
}
