% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_misclass_matrix_01.R
\name{interrogate}
\alias{interrogate}
\title{Accessory function to create_misclass_matrix_01 that helps with trinarisation}
\usage{
interrogate(x, summarymatrix)
}
\description{
Accessory function to create_misclass_matrix_01 that helps with trinarisation
}
