% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kang2017}
\alias{Kang2017}
\title{Count matrices of 8 pooled lupus patient samples within two conditions}
\format{
A list with 5 items:
\describe{
  \item{count_ctrl}{the count matrix for three days after H.polygyrus infection}
  \item{count_stim}{the count matrix for ten days after H.polygyrus infection}
  \item{svm_mat}{the simularity matrix}
  \item{svmDF}{the data frame used to calculate the similarity matrix.}
  \item{source}{the source of this dataset}
}
}
\source{
\url{https://www.nature.com/articles/nbt.4042}
}
\usage{
Kang2017
}
\description{
A data containing the count matrices, the similarity matrix and other 
variables used to generate the similarity matrix 
from single cell RNA sequencing data of 8 pooled lupus patient samples within 
two conditions
}
\keyword{datasets}
