% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_transition.R
\name{KNN_transition}
\alias{KNN_transition}
\title{Calculate stochastic transition matrix between clusters from a KNN
connection matrix}
\usage{
KNN_transition(KNN_matrix, clusters)
}
\arguments{
\item{KNN_matrix}{a sparse binary matrix with size (n_sample, n_sample).
x_ij=1 means sample j is a neighbour of sample i. As definition, we expect
sum(KNN_matrix) = n_sample * K, where K is the number neighbours.}

\item{clusters}{a (n_sample, ) vector of cluster id for each sample.}
}
\description{
The transition probability from cluster i to j is the fraction of neighbours
of all samples in cluster i that belongs to cluster j. Note, this matrix is
asymmetric, so as the input KNN connection matrix.
}
