% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_misclass_matrix_01.R
\name{misclassification_01}
\alias{misclassification_01}
\title{Accessory function to create_misclass_matrix_01 that produces the misclassification matrix}
\usage{
misclassification_01(trinarisedM, clust)
}
\description{
Accessory function to create_misclass_matrix_01 that produces the misclassification matrix
}
