% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DITASIC.R
\name{optim.fit}
\alias{optim.fit}
\title{Accessory functions to estimate count}
\usage{
optim.fit(y, z, start = rep(mean(y), ncol(z)), ...)
}
\arguments{
\item{y, }{please specify}

\item{z, }{please specify}

\item{start, }{please specify}
}
\value{
please specify
}
\description{
Accessory functions to estimate count
}
