% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ren2021}
\alias{Ren2021}
\title{Count matrix and metadata of a large COVID-19 scRNA-seq data cohort.}
\format{
A list with 7 items:
\describe{
  \item{countM}{the count matrix for all samples comming from different condition}
  \item{designM}{the corresponding metadata related to each sample}
  \item{source}{the source of this dataset}
}
}
\source{
\url{https://www.nature.com/articles/nature24489}
}
\usage{
Ren2021
}
\description{
A data containing the count matrix, metadata from a large COVID-19 cohort.
Count matrices are 
calculated based on the number of cells in each cell type. The information 
in the design matrix is collected in the original paper.
}
\keyword{datasets}
