% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{renamecountmatrix}
\alias{renamecountmatrix}
\title{Accessory function to simulate_data that renames the columns of the gene expression count 
matrices of subjects by the order of cell and the index of the subject}
\usage{
renamecountmatrix(x)
}
\description{
Accessory function to simulate_data that renames the columns of the gene expression count 
matrices of subjects by the order of cell and the index of the subject
}
