% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Main simulate data function}
\usage{
simulate_data(x, y, replicates_number, subject_cellnumber)
}
\arguments{
\item{x}{A vector of proportions signifying the desired proportions of cell types for
subjects drawn from a hypothetical condition A. The number of clusters created by the function 
is equal to the length of x}

\item{y}{A vector of proportions signifying the desired proportions of cell types for 
subjects drawn from a hypothetical condition B.}

\item{replicates_number}{An integer signifying the number of subjects drawn from each condition}

\item{subject_cellnumber}{An integer signifying the number of cells drawn from a subject}
}
\value{
a list containing \code{primary_count_matrix}, a large gene expression count matrix simulated from
splatter, containing equal proportions of cells from different clusters; \code{subjects}, a list of
gene expression profiles of different subjects; \code{cluster_lists}, a list of gene expression profiles of cells from different clusters.
}
\description{
Main simulate data function
}
