% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{Simulated dataset with two conditions}
\format{
A list with 5 items:
\describe{
  \item{numb_cond1}{the count matrix of condition 1}
  \item{numb_cond2}{the count matrix of condition 2}
  \item{knn_mat}{the similariy matrix}
  \item{knnGraphs}{the knn graphs information used to calculate the 
  similarity matrix.}
  \item{labels}{the clusters' label for each simulated single cell}
}
}
\usage{
simulation
}
\description{
A data containing the count matrices, the similarity matrix and other 
variables used to generate the similarity matrix 
from a simulated single cell RNA sequencing data with two conditions. 
Dirichlet distribution was used to generate a proportion vector for cell 
types based on the defined true proportions. Multinomial distribution was 
used to generate simulated cell numbers. Cells were selected from cell pools
based on cell numbers and gene expression matrices were processed by Seurat.
The count matrix were calculated by the clustering results, and the 
similarity matrix was calculated using knn graph. The labels are the groud
true annotation of each cell.
}
\keyword{datasets}
