% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatbySeurat.R
\name{treatbySeurat}
\alias{treatbySeurat}
\title{Main function to filter and normalise the large data matrix}
\usage{
treatbySeurat(tip.data)
}
\arguments{
\item{tip.data}{A large data matrix that may be unfiltered and not normalised. Preferably this should be imported with
Read10X() function available from Seurat}
}
\value{
A list containing the filtered and normalised count matrix and the processed seurat object
}
\description{
Main function to filter and normalise the large data matrix
}
