% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similarity_mat.R
\name{get_similarity_mat}
\alias{get_similarity_mat}
\title{Generate similarity matrix with uniform confusion rate to none-self clusters}
\usage{
get_similarity_mat(K, confuse_rate)
}
\arguments{
\item{K}{A integer for number of cluster}

\item{confuse_rate}{A float for confusion rate, uniformly to none-self
clusters}
}
\value{
a similarity matrix with uniform confusion with other cluster
}
\description{
Create a similarity matrix assuming the misclassification error distribute
uniformly in all clusters
}
\examples{
get_similarity_mat(4, 0.1)

}
